requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
spacy==3.5.3

# ── Document processing ────────────────────────────────────────────────────
pypdf==4.2.0
tiktoken==0.7.0

# ── Embeddings & retrieval ─────────────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2

# ── Numerics ───────────────────────────────────────────────────────────────
numpy==1.26.4

# ── UI ─────────────────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.32.3
huggingface_hub==0.16.4
httpx==0.27.0
dotenv==1.0.0